Refactor state store
